move test under src
