+++ src/test/TestMurmurHash.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD-3-clause license. See LICENSE file in the project root for license information.
  */
 
-import { MurmurHash } from '../src/misc/MurmurHash';
+import { MurmurHash } from '../misc/MurmurHash';
 
 import { suite, test } from 'mocha-typescript';
