+++ src/test/tool/TestBufferedTokenStream.tssoon
