+++ src/test/tool/TestCommonTokenStream.tssoon
