+++ src/test/tool/TestIntervalSet.ts
@@ -7,9 +7,9 @@
 // ConvertTo-TS run at 2016-10-04T11:27:11.6934341-07:00
 
 require('source-map-support').install();
-import { IntervalSet } from '../../src/misc/IntervalSet';
-import { Lexer } from '../../src/Lexer';
-import { Token } from '../../src/Token';
+import { IntervalSet } from '../../misc/IntervalSet';
+import { Lexer } from '../../Lexer';
+import { Token } from '../../Token';
 
 import { suite, test as Test } from 'mocha-typescript';
 import * as assert from "assert";
