+++ src/test/tool/TestSymbolIssues.tssoon
