+++ src/test/tool/TestTokenStreamRewriter.tssoon
