+++ src/test/tool/TestTokenTypeAssignment.tssoon
