+++ src/test/tool/TestVocabulary.tssoon
